requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   groq
    #   httpx
build==1.3.0
    # via pip-tools
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
distro==1.9.0
    # via groq
greenlet==3.3.0
    # via sqlalchemy
groq==0.37.1
    # via langchain-groq
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   groq
    #   langgraph-sdk
    #   langsmith
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.2.0
    # via -r requirements.in
langchain-core==1.2.2
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-groq
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-groq==1.1.1
    # via -r requirements.in
langgraph==1.0.5
    # via langchain
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.5
    # via langgraph
langgraph-sdk==0.3.0
    # via langgraph
langsmith==0.5.0
    # via langchain-core
mysql-connector-python==9.5.0
    # via -r requirements.in
numpy==2.3.5
    # via pandas
orjson==3.11.5
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.1
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   build
    #   langchain-core
    #   langsmith
pandas==2.3.3
    # via -r requirements.in
pip-tools==7.5.2
    # via -r requirements.in
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.5
    # via
    #   groq
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
pydantic-core==2.41.5
    # via pydantic
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via langchain-core
requests==2.32.5
    # via
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via groq
sqlalchemy==2.0.45
    # via -r requirements.in
tenacity==9.1.2
    # via langchain-core
typing-extensions==4.15.0
    # via
    #   groq
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.2
    # via requests
uuid-utils==0.12.0
    # via
    #   langchain-core
    #   langsmith
wheel==0.45.1
    # via pip-tools
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
